fixed interval saving
git-svn-id: svn://svn.code.sf.net/p/qgifer/code/trunk@6 8bb6e544-dc85-452e-9350-54fadf7a98b7

+++ src/gifwidget.cpp
@@ -81,7 +81,7 @@ void GifWidget::save()
      if(!prevFrames.size())
 	  return;
      pause();
-     gif.setDuration((float)intervalBox->value()/1000);
+     gif.setDuration((double)intervalBox->value()/1000);
      gif.save(filename.toStdString().c_str(),
 	  saveEveryBox->isChecked() ? seBox->value() : 1);
 }
